+++ zaitsev.vladimir/T1/geometric_functions.hpp
