+++ zaitsev.vladimir/T1/rectangle.cpp
