+++ zaitsev.vladimir/T1/shapes_i_o.hpp
